+++ src/palm/core/behavior_tree/nodes/leaf/__init__.py
@@ -3,11 +3,13 @@
 from palm.core.behavior_tree.nodes.leaf.action_node import ActionNode
 from palm.core.behavior_tree.nodes.leaf.condition_node import ConditionNode
 from palm.core.behavior_tree.nodes.leaf.interactive_leaf import InteractiveLeaf
+from palm.core.behavior_tree.nodes.leaf.resource_leaf import ResourceLeaf
 from palm.core.behavior_tree.nodes.leaf.transform_leaf import TransformLeaf
 
 __all__ = [
     "ActionNode",
     "ConditionNode",
     "InteractiveLeaf",
+    "ResourceLeaf",
     "TransformLeaf",
 ]